+++ krita/plugins/extensions/dockers/specificcolorselector/kis_color_input.cc
@@ -219,7 +219,7 @@ void KisFloatColorInput::update()
     m_dblNumInput->setMaximum(m_maxValue);
 
     // ensure at least 3 significant digits are always shown
-    int newPrecision = 2 + qMax(0.0, std::ceil(-std::log10(m_maxValue)));
+    int newPrecision = 2 + qMax(qreal(0.0), std::ceil(-std::log10(m_maxValue)));
     if (newPrecision != m_dblNumInput->decimals()) {
         m_dblNumInput->setDecimals(newPrecision);
         m_dblNumInput->updateGeometry();
